Project/requirements.txt
geopy==2.2.0
google-api-python-client==2.15.0
joblib==1.0.1
streamlit==0.85.1
streamlit-folium==0.4.0
scikit-learn==0.24.0
python-dotenv==0.19.0
